+++ lockview.js
@@ -1,6 +1,6 @@
 import {registerSettings} from "./src/settings.js";
 import {pushControlButtons} from "./src/controlButtons.js";
-import {registerLayer,Viewbox,drawingConfigApp,closeDrawingConfigApp,getControlledTokens} from "./src/misc.js";
+import {registerLayer,Viewbox,drawingConfigApp,closeDrawingConfigApp,getControlledTokens,blackSidebar,getEnable,getViewboxEnable} from "./src/misc.js";
 import {renderSceneConfig,closeSceneConfig} from "./src/sceneConfig.js";
 import {constrainView_Override,pan_OverrideHigherRes,_Override,pan_Override,onDragCanvasPan_Override,animatePan_Override} from "./src/overrides.js";
 
@@ -18,6 +18,8 @@ let windowWidthOld= -1;
 let windowHeightOld = -1;
 let fitScale = -1;
 export var scaleMax = CONFIG.Canvas.maxZoom;
+
+//CONFIG.debug.hooks = true;
 
 Hooks.on('ready', ()=>{
   game.socket.on(`module.LockView`, (payload) =>{
@@ -45,7 +47,7 @@ Hooks.on('ready', ()=>{
           canvas.stage.addChild(viewbox[senderNumber]);
           viewbox[senderNumber].init();
         }
-
+        
         //update viewbox
         viewbox[senderNumber].updateViewbox(
           {
@@ -59,11 +61,11 @@ Hooks.on('ready', ()=>{
         //viewbox[senderNumber].show;
       }
     }
-    else if (payload.msgType == "lockView_getData" && (game.settings.get("LockView","Enable") || (game.settings.get("LockView","ForceEnable") && game.user.isGM == false))){
+    else if (payload.msgType == "lockView_getData"){
       sendViewBox(payload.senderId);
     }
     else if (payload.msgType == "lockView_update") {
-      if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
+      if (getEnable(game.userId) == false) return;
       let initX = -1;
       let initY = -1;
       if (payload.autoScale > 0 && payload.autoScale < 4) scaleToFit(payload.autoScale);
@@ -74,9 +76,13 @@ Hooks.on('ready', ()=>{
       let lockZoom = payload.lockZoom;
       let boundingBox = payload.boundingBox;
       setBlocks(lockPan,lockZoom,boundingBox);
+      
+      //set sidebar background 
+      let blackenSidebar = (canvas.scene.getFlag("LockView","blackenSidebar") && canvas.scene.getFlag("LockView","excludeSidebar") ? true : false);
+      blackSidebar(blackenSidebar);
     }
     else if (payload.msgType == "lockView_resetView"){
-      if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
+      if (getEnable(game.userId) == false) return;
       let initX = canvas.scene.getFlag('LockView', 'initX');
       let initY = canvas.scene.getFlag('LockView', 'initY');
       let autoScale = canvas.scene.getFlag('LockView', 'autoScale');
@@ -137,6 +143,14 @@ Hooks.on('ready', ()=>{
       setBlocks(lockPan,lockZoom,boundingBox);
     }
     else if (payload.msgType == "lockView_forceCanvasPan") forceCanvasPan();
+    else if (payload.msgType == "lockView_refreshSettings") {
+      if (getEnable(game.userId)) updateSettings();
+      else {
+        setBlocks(false,false,false,true);
+        blackSidebar(false);
+      }
+      sendViewBox(payload.senderId);
+    }
   });
 });
 
@@ -155,9 +169,9 @@ Hooks.once('init', function(){
 
 Hooks.on("canvasInit", (canvas) => {
   if (game.user.isGM) {
-    //for (let i=0; i< viewbox.length; i++)
-    //  if (viewbox[i] != undefined)
-    //    viewbox[i].init();
+    for (let i=0; i< viewbox.length; i++)
+      if (viewbox[i] != undefined)
+        viewbox[i].hide();
   }
 });
   
@@ -190,7 +204,7 @@ Hooks.on('canvasReady',(canvas)=>{
     }
     
   }
-  else if ((game.settings.get("LockView","Enable") || game.settings.get("LockView","ForceEnable"))){
+  else if (getEnable(game.userId)){
     if (canvas.scene.getFlag('LockView', 'lockPan')){
       if (canvas.scene.data.flags["LockView"].initX){}
       else {
@@ -215,20 +229,22 @@ Hooks.on('canvasReady',(canvas)=>{
   updateSettings();
   checkKeys();
   forceCanvasPan();
+  sendViewBox(game.data.users.find(users => users.role == 4)._id);
 });
 
 Hooks.on('canvasPan',(canvas,data)=>{
-  if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
-  scaleMax = CONFIG.Canvas.maxZoom;
-  scaleToFit();
-  let autoScale = canvas.scene.getFlag('LockView', 'autoScale');
-  if (fitScale > 0 && fitScale != canvas.scene._viewPosition.scale && (autoScale > 0 && autoScale < 4)){
-    if (Math.abs(canvas.scene._viewPosition.scale - fitScale) < 0.015){
-      Canvas.prototype.pan = pan_OverrideHigherRes;
-      updateView(-1,-1,fitScale);
-      let lockPan = canvas.scene.getFlag('LockView', 'lockPan');
-      if (lockPan) Canvas.prototype.pan = pan_Override;
-      else if (canvas.scene.getFlag('LockView', 'boundingBox')==false) Canvas.prototype.pan = pan_Default;
+  if (getEnable(game.userId)) {
+    scaleMax = CONFIG.Canvas.maxZoom;
+    scaleToFit();
+    let autoScale = canvas.scene.getFlag('LockView', 'autoScale');
+    if (fitScale > 0 && fitScale != canvas.scene._viewPosition.scale && (autoScale > 0 && autoScale < 4)){
+      if (Math.abs(canvas.scene._viewPosition.scale - fitScale) < 0.015){
+        Canvas.prototype.pan = pan_OverrideHigherRes;
+        updateView(-1,-1,fitScale);
+        let lockPan = canvas.scene.getFlag('LockView', 'lockPan');
+        if (lockPan) Canvas.prototype.pan = pan_Override;
+        else if (canvas.scene.getFlag('LockView', 'boundingBox')==false) Canvas.prototype.pan = pan_Default;
+      }
     }
   }
   sendViewBox(game.data.users.find(users => users.role == 4)._id,data);
@@ -241,7 +257,7 @@ Hooks.on("getSceneControlButtons", (controls) => {
 });
 
 Hooks.on("updateDrawing",()=>{
-  if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
+  if (getEnable(game.userId) == false) return;
   forceCanvasPan();
 });
 
@@ -296,10 +312,14 @@ Hooks.on("closeSceneConfig", (app, html) => {
   closeSceneConfig(app,html);
 });
 
+Hooks.on("sidebarCollapse", () => {
+  updateSettings();
+})
+
 //This hook is the last hook that is called when initializing scene. It's used to make sure that the payload that's sent is the most recent
 Hooks.on('renderSettings',()=>{
-  if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
-    sendViewBox(game.data.users.find(users => users.role == 4)._id);
+  //sendViewBox(game.data.users.find(users => users.role == 4)._id);
+  if (getEnable(game.userId) == false) return;
   const lockPan = canvas.scene.getFlag('LockView', 'lockPan');
   const lockZoom = canvas.scene.getFlag('LockView', 'lockZoom');
   const boundingBox = canvas.scene.getFlag('LockView', 'boundingBox');
@@ -346,24 +366,31 @@ export function sendLockView_update(lockPan,lockZoom,autoScale,forceInit,boundin
 function scaleToFit(force = 0){
   let horizontal;
   let autoScale = canvas.scene.getFlag('LockView', 'autoScale');
+  let excludeSidebar = false;
+  let sidebarOffset = 0;
+  if (canvas.scene.getFlag("LockView","excludeSidebar") && ui.sidebar._collapsed == false) {
+    excludeSidebar = true;
+    sidebarOffset = window.innerWidth-ui.sidebar._element[0].offsetLeft;
+  }
+
   if ((autoScale == 1 && force == 0) || force == 1) horizontal = true;
   else if ((autoScale == 2 && force == 0) || force == 2) horizontal = false;
   else if ((autoScale == 3 && force == 0) || force == 3) {
-    if ((window.innerWidth / canvas.dimensions.sceneWidth) > (window.innerHeight / canvas.dimensions.sceneHeight)) horizontal = true;
+    if (((window.innerWidth-sidebarOffset) / canvas.dimensions.sceneWidth) > (window.innerHeight / canvas.dimensions.sceneHeight)) horizontal = true;
     else horizontal = false;
   }
   else return;
   Canvas.prototype.pan = pan_OverrideHigherRes;
 
-  let x = canvas.dimensions.paddingX + canvas.dimensions.sceneWidth/2;
-  let y = canvas.dimensions.paddingY + canvas.dimensions.sceneHeight/2;
   let scale = -1;
   if (horizontal){
     let windowWidth = window.innerWidth;
     let sceneWidth = canvas.dimensions.sceneWidth;
     if (windowWidth != windowWidthOld || force > 0){
       windowWidthOld = windowWidth;
-      scale = windowWidth/sceneWidth;
+      scale = (windowWidth-sidebarOffset)/sceneWidth;
+      let x = canvas.dimensions.paddingX + (canvas.dimensions.sceneWidth+sidebarOffset/scale)/2;
+      let y = canvas.dimensions.paddingY + canvas.dimensions.sceneHeight/2;
       fitScale = Math.round(scale* 2000) / 2000;
       updateView(x,y,scale);
     }
@@ -374,6 +401,8 @@ function scaleToFit(force = 0){
     if (windowHeight != windowHeightOld || force > 0){
       windowHeightOld = windowHeight;
       scale = windowHeight/sceneHeight;
+      let x = canvas.dimensions.paddingX + (canvas.dimensions.sceneWidth+sidebarOffset/scale)/2;
+      let y = canvas.dimensions.paddingY + canvas.dimensions.sceneHeight/2;
       fitScale = Math.round(scale* 2000) / 2000;
       updateView(x,y,scale);
     }
@@ -386,7 +415,7 @@ function scaleToFit(force = 0){
 }
 
 export function updateSettings(){
-  if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
+  if (getEnable(game.userId) == false) return;
   let initX = canvas.scene.getFlag('LockView', 'initX');
   let initY = canvas.scene.getFlag('LockView', 'initY');
   let lockPan = canvas.scene.getFlag('LockView', 'lockPan');
@@ -399,6 +428,9 @@ export function updateSettings(){
   else if (autoScale == 4) updateView(-1,-1,getScale());
   else if (forceInit) updateView(initX,initY,canvas.scene.data.initial.scale);
   setBlocks(lockPan,lockZoom,boundingBox);
+  //set sidebar background 
+  let blackenSidebar = (canvas.scene.getFlag("LockView","blackenSidebar") && canvas.scene.getFlag("LockView","excludeSidebar") ? true : false);
+  blackSidebar(blackenSidebar);
 }
 
 function getScale(){
@@ -425,8 +457,21 @@ function updateView(moveX,moveY,scale){
 
 //Send data to the GM to draw the viewbox
 function sendViewBox(target,viewPosition=undefined){
+  if (getViewboxEnable(game.userId)==false) return;
   if (viewPosition == undefined) viewPosition = canvas.scene._viewPosition;
-  
+
+  const excludeSidebar = canvas.scene.getFlag('LockView', 'excludeSidebar');
+  const blackSidebar = canvas.scene.getFlag('LockView', 'blackenSidebar');
+  let viewPositionNew = {
+    x: viewPosition.x,
+    y: viewPosition.y,
+    scale: viewPosition.scale
+  }
+  let offset = 0;
+  if (ui.sidebar._collapsed == false && excludeSidebar && blackSidebar){
+    offset = (window.innerWidth-ui.sidebar._element[0].offsetLeft);
+    viewPositionNew.x -= offset/(2*viewPosition.scale);
+  }
   let payload = {
       "msgType": "lockView_viewport",
       "senderId": game.userId, 
@@ -434,20 +479,21 @@ function sendViewBox(target,viewPosition=undefined){
       "senderColor": game.user.color,
       "receiverId": target, 
       "sceneId": canvas.scene.data._id,
-      "viewPosition": viewPosition,
-      "viewWidth": window.innerWidth,
+      "viewPosition": viewPositionNew,
+      "viewWidth": window.innerWidth-offset,
       "viewHeight": window.innerHeight
   };
+
   game.socket.emit(`module.LockView`, payload);
 }
 
 //Block zooming and/or panning
-function setBlocks(lockPan,lockZoom,boundingBox){
-  if (game.settings.get("LockView","Enable") == false && (game.settings.get("LockView","ForceEnable") == false || game.user.isGM)) return;
+function setBlocks(lockPan,lockZoom,boundingBox,force=false){
+  if (force==false && getEnable(game.userId) == false) return;
 
   if (lockZoom == true) Canvas.prototype._onMouseWheel = _Override;
   else if (lockZoom == false) Canvas.prototype._onMouseWheel = _onMouseWheel_Default;
-  
+
   if (lockPan) {
     Canvas.prototype._onDragCanvasPan = _Override;  //draggin token to edge of screen
     Canvas.prototype.pan = pan_Override;  //right-mouse click & zoom
@@ -467,6 +513,7 @@ function setBlocks(lockPan,lockZoom,boundingBox){
 }
 
 function checkKeys() {
+  /*
   let fired = false;
   
   window.addEventListener("keydown", async (e) => {
@@ -476,16 +523,7 @@ function checkKeys() {
       window.Azzu.SettingsTypes.KeyBinding.parse(game.settings.get('LockView','lockOverride')))
   ) {
       fired = true;
-      let lockPan = canvas.scene.getFlag('LockView', 'lockPan');
-      let lockZoom = canvas.scene.getFlag('LockView', 'lockZoom');
-      let boundingBox = canvas.scene.getFlag('LockView', 'boundingBox');
-      lockPanStorage = lockPan;
-      lockPan = false;
-      lockZoomStorage = lockZoom;
-      lockZoom = false;
-      boundingBoxStorage = boundingBox;
-      boundingBox = false;
-      setBlocks(lockPan,lockZoom,boundingBox);
+      setBlocks(false,false,false);
     }
   });
 
@@ -494,7 +532,7 @@ function checkKeys() {
     const lockPan = canvas.scene.getFlag('LockView', 'lockPan');
     const lockZoom = canvas.scene.getFlag('LockView', 'lockZoom');
     const boundingBox = canvas.scene.getFlag('LockView', 'boundingBox');
-
     setBlocks(lockPan,lockZoom,boundingBox);
   });
+  */
 }