+++ src/sceneConfig.js
@@ -6,6 +6,8 @@ export function renderSceneConfig(app,html){
     let autoScale = 0;
     let forceInit = false;
     let boundingBox = false;
+    let excludeSidebar = false;
+    let blackenSidebar = false;
     if(app.object.data.flags["LockView"]){
         if (app.object.data.flags["LockView"].lockPan_Default){
         lockPan_Default = app.object.getFlag('LockView', 'lockPan_Default');
@@ -28,6 +30,14 @@ export function renderSceneConfig(app,html){
         if (app.object.data.flags["LockView"].boundingBox){
             boundingBox = app.object.getFlag('LockView', 'boundingBox');
         } else app.object.setFlag('LockView', 'boundingBox', false);
+
+        if (app.object.data.flags["LockView"].excludeSidebar){
+            excludeSidebar = app.object.getFlag('LockView', 'excludeSidebar');
+        } else app.object.setFlag('LockView', 'excludeSidebar', false);
+
+        if (app.object.data.flags["LockView"].blackenSidebar){
+            blackenSidebar = app.object.getFlag('LockView', 'blackenSidebar');
+        } else app.object.setFlag('LockView', 'blackenSidebar', false);
     } 
     let autoScaleOptions = [
         game.i18n.localize("LockView.Scene.Autoscale.Off"),
@@ -46,6 +56,9 @@ export function renderSceneConfig(app,html){
     autoScaleSelected[autoScale] = "selected"
 
     const fxHtml = `
+    <header class="form-header">
+        <h3><i class="fas fa-lock"/></i> Lock View</h3>
+    </header>
     <div class="form-group">
         <label>${game.i18n.localize("LockView.Scene.LockPan")}</label>
         <input id="LockView_lockPan" type="checkbox" name="LV_lockPan" data-dtype="Boolean" ${lockPan_Default ? 'checked' : ''}>
@@ -73,14 +86,23 @@ export function renderSceneConfig(app,html){
         <p class="notes">${game.i18n.localize("LockView.Scene.Autoscale_Hint")}</p>
     </div>
     <div class="form-group">
+        <label>${game.i18n.localize("LockView.Scene.ExcludeSidebar")}</label>
+        <input id="LockView_excludeSidebar" type="checkbox" name="LV_excludeSidebar" data-dtype="Boolean" ${excludeSidebar ? 'checked' : ''}>
+        <p class="notes">${game.i18n.localize("LockView.Scene.excludeSidebar_Hint")}</p>
+    </div>
+    <div class="form-group">
+        <label>${game.i18n.localize("LockView.Scene.BlackenSidebar")}</label>
+        <input id="LockView_blackenSidebar" type="checkbox" name="LV_blackenSidebar" data-dtype="Boolean" ${blackenSidebar ? 'checked' : ''}>
+        <p class="notes">${game.i18n.localize("LockView.Scene.blackenSidebar_Hint")}</p>
+    </div>
+    <div class="form-group">
         <label>${game.i18n.localize("LockView.Scene.ForceInit")}</label>
         <input id="LockView_forceInit" type="checkbox" name="LV_forceInit" data-dtype="Boolean" ${forceInit ? 'checked' : ''}>
         <p class="notes">${game.i18n.localize("LockView.Scene.ForceInit_Hint")}</p>
     </div>
     `
-    const fxFind = html.find("input[name ='initial.x']");
-    const formGroup = fxFind.closest(".form-group");
-    formGroup.after(fxHtml);
+    const submitBtn = html.find("button[name = 'submit']");
+    submitBtn.before(fxHtml);
 }
 
 export function closeSceneConfig(app,html){let lockPan = html.find("input[name ='LV_lockPan']").is(":checked");
@@ -88,6 +110,8 @@ export function closeSceneConfig(app,html){let lockPan = html.find("input[name =
     let autoScale = html.find("select[name='LV_autoScale']")[0].value;
     let forceInit = html.find("input[name ='LV_forceInit']").is(":checked");
     let boundingBox = html.find("input[name ='LV_boundingBox']").is(":checked");
+    let excludeSidebar = html.find("input[name ='LV_excludeSidebar']").is(":checked");
+    let blackenSidebar = html.find("input[name ='LV_blackenSidebar']").is(":checked");
 
     app.object.setFlag('LockView', 'lockPan',lockPan);
     app.object.setFlag('LockView', 'lockZoom',lockZoom);
@@ -96,6 +120,8 @@ export function closeSceneConfig(app,html){let lockPan = html.find("input[name =
     app.object.setFlag('LockView', 'autoScale',autoScale);
     app.object.setFlag('LockView', 'forceInit', forceInit);
     app.object.setFlag('LockView', 'boundingBox', boundingBox);
+    app.object.setFlag('LockView', 'excludeSidebar', excludeSidebar);
+    app.object.setFlag('LockView', 'blackenSidebar', blackenSidebar);
 
     if (app.entity.data._id == canvas.scene.data._id){
         let initX,initY;